Add glossary.md to docs
Motivation
We should have a page in our documentation reserved to defining parts of
our framework.

Modifications
I've created a file named glossary with the definition of a UI Component
and the definition of a RootViewController.

Result
A new file in our docs folder has been added responsible for bein the
glossary of our documentation.